Parsing config and determining OS
Added code to parse the config, pull out the variables and then compile
the regex. Then use this and determine the OS.

As we will need to use variables that are pulled from the message prefix
I had passed all of the variables that are defined in the prefix
config to the child process so they can be used later.