Verify config and compile messages regex
Add a check to verify the config. At the moment we are only using dicts
and strings in the config, so only added checks for those. If in future
we add any lists then we will need to update the config checker.

Complile the regex for each OS and store the dict ready to be checked
against. We will check the error title and if it matches then we will
use the regex, this should cut down on regex checks required.