Change transport to use multiprocess Pipe
It is not possible to pass a zeromq socket to a child process as it
seems that it does not have the permission to write to the socket.
Therefore I have changed the transport to use a multiprocessing Pipe
then the socket can be initiated and published to in the same process.