Add static mappings
Previously we only had mappings where the value was pulled stright from
the syslog message. However sometimes we will need to add static
mappings, i.e in the examples that I have added for `junos` and `eos`,
in the `junos` example when a MD5 sum is incorrect we know that the
session is not established so can set it accordingly.

I have changed the order of the mappings from `{result: oc_mapping}` to
`{oc_mapping: result}` as the result does not need to be unique, but the
mapping does, with the result as the key if two mappings had the same
key, one would be overwritten.

I also changes the config check to allow for an empty dict.