Add replace functionality to manipulate a syslog variable
The adds the ability to replace / transform a variable that is pulled
from the syslog message.

The example  that caused this to be added is `SNMP_TRAP_LINK_DOWN` for
`junos`. The openconfig model expects the value to be uppercase, but the
syslog message presents it in lowercase, therefore we need to be able to
replace `down` with `DOWN`.

I also had to change the delimeter as we had to set the interface name
within the mapping, and the interface name contained `/`