Check that the cert and key are present and valid issue #50
Current the napalm-logs process will start correctly even if the cert
and key are either not present or do not match. Then it will error when
a client tries to subscribe.

This PR adds a check when starting to make sure the files are present
and that the key and cert match.