Catch SIGINT (Ctrl-c) and exit cleanly
I am adding signal to catch `SIGINT` in `cli.py` and not send it to the
child processes. When a `SIGINT` is caught, `cli` then calls the
`stop_engine` funcion in `napalm_log`, which sends a SIGTERM to each
child process. The child process will catch this `SIGTERM` and run it's
own `stop` function, while will close all pipes / sockets, and terminate
the loop.