Update IOS-XR prefix to include host and node-id
The format of an IOS-XR syslog message is as follows:

```
node-id : timestamp : process-name [pid] : % message category -group
-severity -message -code : message-text
```

The node-id is of the form `RP/0/RP0/CPU0`, and doesn't indicate the
actual host that the message comes from. The only way that I can find to
get the hostname in there is by using a hostname prefix:

```
logging hostnameprefix <hostname>
```

This then produces a message as follows:

```
<149>12345678: edge01.aaa01 RP/0/RSP1/CPU0:Mar 28 15:08:30.941 UTC:
bgp[1051]: %ROUTING-BGP-5-MAXPFX : No. of IPv4 Unicast prefixes received
from 1.2.3.4 has reached 100, max 110
```