Add v6 match for junos BGP_PREFIX_THRESH_EXCEEDED
The OC model requires you to specify the `afi-safi-name`, which for ipv4
unicast is `ipv4-unicast`, and for ipv6 unicast is `ipv6-unicast`,
because `ipv4` or `ipv6` are not specified in the log message we are not
able to determine this with our current config. Therefore I am adding a
new error which will match the ipv6 address, and has `ipv6-` hardcoded
in the oc model.