Add original message if unable to identify OS and parse message
Currently if we are unable to identify the OS we send an empty dict to
syslog out. This PR changes it so it sends the original message.

It is also changing the way we send address and port to the logger
transport as you cannot send an argument of the same name twice, and as
the address / port are set in the config file for logger it was doing
so.