Add error BGP_PEER_NOT_CONFIGURED
The format of this error for `junos` is a little different from others:

```
<87>Jul  5 05:52:44  re0.edge01.lhr01 rpd[1848]: bgp_read_message:2764:
NOTIFICATION received from 1.2.3.4 (External AS 1234): code 6 (Cease)
subcode 5 (Connection Rejected)
```

Where you would normally find the tag, we have `bgp_read_message:2764`,
or `2764` could be part of the message, either way it is not standard.

Therefore I have changed the prefix config to not have the whitespace at
the end of the prefix `line`. Along with this I have removed any
whitespace from the beggening (and end) of the message line as most
messages would have a space at the start.