Add time format for all OS
I have added a `time_format` key to the perfix of all `yml` prefix
files.

For the py profiler prefixes I have added a new function called
`time_format` which will just return the time format. I am not sure if
this is the best way, but didn't want to just attach it to the regex
dict.