Add Makefile and update docker config
Rather than having a config file and setting each variable via an env
variable I think it is cleaner to mount the local config file stright
into the docker container, this way any part of the config can change
without having to rebuild the image.

I have added `make run` and `make run-foreground`, both of which will
find the appropriate ports from the config file and proxy these. As it
is not easy to figure out within the `Makefile` if the port is `tcp` or
`udp` I have proxied all ports for both.