Add check to see if timestamp is in future
This goes part way to fixing #204

Currently if the message that we receive does not contain the year we
assume that it is from the current year. This will only really cause an
issue (other than the failing tests) when we are changing years; if a
log is generated at 23:59:59 on 31st dec and received a second later
than the timestamp will be almost a year out.