Remote timestamp key before testing yang outputs
Most syslog messages do not contain a year. This means that
`napalm-logs` has to guess the year, most of the time it will use the
current year. This means that any tests added over a year ago will fail
due to the timestamps not matching. We could manipulate the timestamp in
the test code, but I see no point in doing this as we would need to use
a similar logic in the test code as we do in the actual code to generate
the expected timestamp, which is pointless. The other alternative is to
update the timestamps every year, but this doesn't sound very sensible.