Allow message to match on multiple OS
This fixes #242

Currently when a matching OS is found we do not check remaining OS. This
mean that if 2 OS have message prefixes that match each other's logs
then you cannot be sure that the message will be forwarded onto the
correct `device` process.

I have updated so the message is now forwarded onto all matching os, not
just the first match.

To allow the tests to work correctly I have removed `'send_raw': True`
from the config test.