Expose publish metrics with more attrbutes
Adding another metrics ``napalm_logs_device_published_messages_attrs``
besides the existing ``napalm_logs_device_published_messages`` to
publish metrics with more attributes. As this may potentially generate
a large amount of metrics I opted not to replace the existing one, and
rather have it separately, and with the possibility to disable it in
case it causes troubles.