Add log processing delay metric
Expose a new metric ``napalm_logs_processing_delay`` that provides
a (rough?) view on how long it takes a message to be processed through
napalm-logs - from the moment the message is received, till published.
This timing excludes the time spent effectively publishing (as it may
depend of various external factors), but it includes the time spent in
the internal queues.