Update PROCESS_CEILING_WATERMARK_REACHED error processing
`jlaunchd` doesn't present its messages with a tag. As such, with the default parsing, the `jlaunchd` message interpreted `System reaching processes ceiling high watermark` as the tag. Using this would mean duplication of the syslog parsing, as we'd have to duplicate for every instance of severity.
Updated the `init.yml` file to add an explicit parser for `jlaunchd` messages where `jlaunchd` becomes the tag.
Added a test as well