Add a simplified way to have counter-part messages
From the YAML profiles, which can be easily extended and loaded into the
user environment, we should be able to signalise groups or pairs of
notifications, referring to the same entity (i.e., interface, neighbour,
alarm, etc.). This would have a direct applicability into the Prometheus
metrics, as it'd add a new `_state` metric by default, but also would be
useful in the other publishers which will be adding the new ``state``
field when provided. The new ``state`` field is a numeric value, with
a specific meaning, mapping out to a value that provides a meaningful
information for the end user (e.g., interface up = 1, interface down = 0;
alarm set = 1, alarm cleared = 0, and so on.) - but not limited to
binary values only, there can be as many states as desired.