[385] Fix for stuck consumers due to unsafe host key access
Vendors like IOS-XR by default send messages with no hostname. Log parsers were updated to support this.

This PR fixes the server.py unsafe access for the host key in msg_dict which cause the consumer/listeners to become stuck
as they are expecting messages to have a host key for lookup.

Signed-off-by: Adrian Arumugam <adrian@a6m.au>